lib/rpmdb.c dbiFindMatches(): make the types to pass the input value match its source
Since we read epoch with strtoll(), it's known to be long long.

long long is a basic type and is at least 64-bit. (And splitEpoch()
checks that it's in the range from 0 to UINT32_MAX.)

It's cleaner if we pass it in a variable whose type is called long
long everywhere (instead of int64_t), and then compare it to whatever
integer value we get from a query (uint64_t from rpmtdGetNumber();
note that it originates from a 32-bit value actually).

We don't use an unsigned type here because the convention is to pass
an absent value as -1. If the convention was different, we could use
unsigned long long (or even unsigned long, because it's good to
holding values up to UINT32_MAX.)

+++ lib/rpmdb.c
@@ -113,9 +113,9 @@ static unsigned int uintId(unsigned int a)
  * @return		epoch integer within the [0; UINT32_MAX] interval,
  *                      or -1 for no epoch
  */
-static int64_t splitEpoch(const char *s, const char **version)
+static long long splitEpoch(const char *s, const char **version)
 {
-    int64_t e;
+    long long e;
     char *end;
     int saveerrno = errno;
 
@@ -805,7 +805,7 @@ int rpmdbCountPackages(rpmdb db, const char * name)
  */
 static rpmRC dbiFindMatches(rpmdb db, dbiIndex dbi,
 		const char * name,
-		int64_t epoch,
+		long long epoch,
 		const char * version,
 		const char * release,
 		const char * arch,
@@ -902,7 +902,7 @@ static rpmRC dbiFindByLabelArch(rpmdb db, dbiIndex dbi,
 			    dbiIndexSet * matches)
 {
     char localarg[arglen+1];
-    int64_t epoch;
+    long long epoch;
     const char * version;
     const char * release;
     char * s;
